requirements.txt
openai
chromadb
python-dotenv
langchain-text-splitters
pypdf